# remote write: simplify readability of timeseries filtering by using the slices package