Hopefully this will help us fix #13979 or reproduce it in a test more easily. It's been noted a few times in https://github.com/prometheus/prometheus/pull/14078 that the readability or modification of the underlying `pendingData` slice is difficult and potentially the cause of the data corruption.

cc @FUSAKLA @david-vavra

Without this change
```
perflock go test -run=XXX -bench=BenchmarkBuildTimeSeries -benchmem -timeout=99999s
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: AMD Ryzen 9 5950X 16-Core Processor
BenchmarkBuildTimeSeries-32    	     529	   2305662 ns/op	 2652227 B/op	   39749 allocs/op
PASS
ok  	github.com/prometheus/prometheus/storage/remote	1.460s
```

With this change
```
perflock go test -run=XXX -bench=BenchmarkBuildTimeSeries -benchmem -timeout=99999s
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: AMD Ryzen 9 5950X 16-Core Processor
BenchmarkBuildTimeSeries-32    	     523	   2355820 ns/op	 2652200 B/op	   39749 allocs/op
PASS
ok  	github.com/prometheus/prometheus/storage/remote	1.475s
```