@bboreham 2025-01-07
Hello from the bug-scrub!
Seems that the underlying issue was fixed in #14078, and the benchmark was added.
So can this PR be simplified?