@cstyan 2025-08-01
Benchmark diff is minimal still:
```
benchstat old new
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote
cpu: AMD Ryzen 7 9800X3D 8-Core Processor
                   │     old     │                new                 │
                   │   sec/op    │   sec/op     vs base               │
BuildTimeSeries-16   174.3µ ± 1%   170.4µ ± 1%  -2.21% (p=0.000 n=10)

                   │    old     │                new                │
                   │    B/op    │    B/op     vs base               │
BuildTimeSeries-16   401.0 ± 1%   394.5 ± 2%  -1.62% (p=0.037 n=10)

                   │    old     │              new              │
                   │ allocs/op  │  allocs/op   vs base          │
BuildTimeSeries-16   5.500 ± 9%   5.000 ± 20%  ~ (p=0.350 n=10)
```

Note that also the benchmark was not correct before, we were accounting for the building of the actual proto in the benchmark timings.